+++ lib/osm.cc
@@ -1,4 +1,3 @@
-#define PYBIND11_DETAILED_ERROR_MESSAGES
 #include <pybind11/pybind11.h>
 #include <pybind11/operators.h>
 
@@ -6,7 +5,6 @@
 
 #include "cast.h"
 #include "osm_base_objects.h"
-#include "osm_helper.h"
 
 namespace py = pybind11;
