+++ lib/osmium.cc
@@ -1,4 +1,3 @@
-#define PYBIND11_DETAILED_ERROR_MESSAGES
 #include <pybind11/pybind11.h>
 
 #include <osmium/osm.hpp>
