+++ lib/simple_writer.cc
@@ -1,4 +1,3 @@
-#define PYBIND11_DETAILED_ERROR_MESSAGES
 #include <pybind11/pybind11.h>
 
 #include <osmium/osm.hpp>
@@ -10,7 +9,6 @@
 
 #include "cast.h"
 #include "osm_base_objects.h"
-#include "osm_helper.h"
 
 namespace py = pybind11;
